+++ src/Platform/Linux/System/Dispatcher.h
@@ -108,7 +108,7 @@ private:
   NativeContext* firstResumingContext;
   NativeContext* lastResumingContext;
   NativeContext* firstReusableContext;
-  size_t runningContextCount;
+  uint64_t runningContextCount;
 
   void contextProcedure(void* ucontext);
   static void contextProcedureStatic(void* context);
